Refactor: orchestrator modularization + A2A hardening
Code review fixes:

1. Split orchestrator.py into subpackage modules:
   - capabilities.py: dynamic MCP/agents/skills management
   - mcp_loader.py: MCP config parsing
   - prompts.py: system prompt builder
   - hooks.py: pre/post tool hooks
   - core.py: main orchestrator (1627→1114 lines)

2. A2ATaskStore: add connection pooling with threading.Lock,
   single WAL connection, 30s timeout

3. Executor: add configurable timeout (default 300s) for blocking
   tasks with proper TimeoutError handling

4. Exception handling: log silent exceptions in task_store,
   handle CancelledError in streaming, specific exceptions in auth
